Update the ListView with task entered by user
I had a problem where I was not able to fetch the result of the database
insertion from the inside of the async block. Because SeaORM needed
async but Cursive did not. I needed id, so I changed the create_task
method. I made the method need id parameter to be provided manually.

This way I was able to have id with with and not lose it inside the
async block of tokio::spawn.

This code may not be entirely fail-proof. For example, in case of a
situation where insertion fails, but I have updated the UI.